requirements.txt
flask==3.0.0
line-bot-sdk==3.11.0
python-dotenv==1.0.1
pytest==8.3.3
google-auth==2.36.0
